### 2025-09-30 #9104 @kyleniemeyer
@marchdf it looks like the paper issue is due to an affiliation block needing quotation marks:
```
affiliation: "1, 2"
```
