### 2025-09-30 #9104 @marchdf
@kyleniemeyer I think I fixed the DOIs and the "generate pdf" step.

Re the articles above. The ERF, MAESTROeX, and chombo-discharge are relevant in that they are also AMReX based codes. The AMReX reference is very relevant because this is an important library that MARBLES leverages. None of the codes referenced implement the lattice Boltzmann method though. The last reference, Coral, is an interesting PDE solver but also does not implement the lattice Boltzmann method.
