### 2025-10-06 #9104 @kyleniemeyer
@marchdf paper looks good (in terms of compiling now), so I will let @philipcardiff take things over and next find reviewers for your submission
