### 2025-11-29 #9530 @editorialbot
[REVIEW]: 'MARBLES: Multi-scale Adaptively Refined Boltzmann LatticE Solver
**Submitting author:** @marchdf (<a href="http://orcid.org/0000-0001-7742-1565">Marc T. Henry de Frahan</a>)
**Repository:** https://github.com/NREL/marbles
**Branch with paper.md** (empty if default branch): add-paper
**Version:** v0.2.0
