### 2025-12-10 #9530 @philipcardiff
> After reading the COI policy, I realized that I have COIs to declare. I apologize for not recognizing this earlier. I had assumed it was acceptable since I was invited to review.
>
> I am a co-author with [@marchdf](https://github.com/marchdf) and [@jrood-nrel](https://github.com/jrood-nrel) on several recent publications:
>
> * https://journals.sagepub.com/doi/full/10.1177/10943420221121151
> * https://joss.theoj.org/papers/10.21105/joss.05450
> * https://epubs.siam.org/doi/abs/10.1137/1.9781611977967.2
>
> I am also a co-author with [@marchdf](https://github.com/marchdf) on the following paper:
>
> * https://www.sciencedirect.com/science/article/abs/pii/S0021999124001281
>
> I think I can still be impartial in reviewing this submission, but I defer to [@philipcardiff](https://github.com/philipcardiff) to make the determination based on this disclosure.

Thanks @WeiqunZhang, for declaring these potential conflicts of interest.

@kyleniemeyer, would you mind letting me know your thoughts?
