### 2025-12-14 NREL/marbles#164 @WeiqunZhang
JOSS paper reference: Duplicated URLs
In the JOSS submission (https://github.com/openjournals/joss-reviews/issues/9530), the URL for de Frahan, Young, Sitaraman, Rood, & Larsen (2023) appears twice. Maybe the `url` in `howpublished` can be removed.
