### 2025-12-14 NREL/marbles#165 @WeiqunZhang
Build instructions for GPU support
The documentation at https://nrel.github.io/marbles/index.html does not currently include build instructions for GPU support. Given that GPU support is one of the main features of this code, it would be helpful to provide guidance on how to build and run MARBLES on GPU platforms.

https://github.com/openjournals/joss-reviews/issues/9530
