### 2025-12-14 NREL/marbles#166 @WeiqunZhang
Missing `example.inp` in `Build/`
According to the instructions at https://nrel.github.io/marbles/index.html#getting-started, there should be a file named `example.inp` under `Build/`.

https://github.com/openjournals/joss-reviews/issues/9530
