### 2026-02-23 natlabrockies/marbles#166 @nileshsawant
I placed a input file in Build folder as part of https://github.com/NatLabRockies/marbles/pull/169 . Thanks @WeiqunZhang
