### 2026-08-16 #9530 @pkkolluru
Hi @philipcardiff I have completed the review of software and accompanying manuscript. I'm happy to recommend accepting both of them in their current form.
